@@ -171,8 +171,8 @@ if (typeof window !== 'undefined' && window.Vue) {
   install(window.Vue);
 }
 
-module.exports = {
-  version: '2.4.8',
+export default {
+  version: '2.5.2',
   locale: locale.use,
   i18n: locale.i18n,
   install,
@@ -247,5 +247,3 @@ module.exports = {
   Main,
   Footer
 };
-
-module.exports.default = module.exports;
