@@ -32,8 +32,7 @@
     props: {
       title: {
         type: String,
-        default: '',
-        required: true
+        default: ''
       },
       description: {
         type: String,
